refactor(theme): move theme-toggle from providers to shared directory
- Relocate theme-toggle.tsx to shared directory as it's a reusable component
- Update import path in header component
